Move spa config to own header file
